Move RoomClass to separate file
Moves the RoomClass and its related files
to a separate script file so that the
FloorManager script only has to maintain its
own logic
