+++ SwipeRogueUnity/Assets/Scripts/CSRoomClass.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 7fc9315f1ca554d05a07854526e526d9
+timeCreated: 1488945606
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
